raspberryPi: mitigate sensor restart problem
On Raspberry Pi 3, if sensor restarts while hub stays on, sometimes
peripheral couldn't discover service and characteristics correctly
and sometimes it stuck on notify enabling. For the latter, we
workaround it by detecting this scenario by a timer and then
disconnect the sensor. The sensor usually works after this. For the
former, we haven't find good way to solve it yet.