workaround DiaperSens immedite connection after turning on problem
DiaperSens starts to send BLE advertisement frames before it's
ready to accept BLE connections. We workaround this probme by
recording the 'first adv' timestamp and holding on connection
until 10s later. The drawback is, if SensEdge restarts it won't
connect to all DiaperSens' around until 10s later. But we should
be able to tolerant this corner case.